requirements.txt
langchain==1.2.10
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-groq==1.1.2
langchain-openai==1.1.10
chromadb==1.5.2
PyMuPDF==1.27.1
gradio==6.8.0
python-dotenv==1.2.2
tiktoken==0.12.0